% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_available_urban}
\alias{check_available_urban}
\title{Get all of the urban areas in the Greenspace Seasonality Data Cube}
\usage{
check_available_urban(test = FALSE)
}
\arguments{
\item{test}{logical. (ignored) Only for testing.}
}
\value{
dataframe
}
\description{
This function returns all of the urban areas
in the Greenspace Seasonality Data Cube dataset.
}
\note{
You can explore all available urban areas in an interacive map at:
\url{https://github.com/billbillbilly/greenSD/blob/main/scripts/city_urban_boundaries.geojson}
}
\examples{
check_available_urban(test = TRUE)
}
\references{
Wu, S., Song, Y., An, J. et al. High-resolution greenspace dynamic
data cube from Sentinel-2 satellites over 1028 global major cities.
Sci Data 11, 909 (2024). https://doi.org/10.1038/s41597-024-03746-7
}
