% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_urban_boundary}
\alias{check_urban_boundary}
\title{Get an urban area boundary based on the UID}
\usage{
check_urban_boundary(uid = NULL, plot = TRUE, test = FALSE)
}
\arguments{
\item{uid}{numeric. Urban area ID. To check the ID of an available urban area,
use \code{\link[=check_available_urban]{check_available_urban()}}}

\item{plot}{logical. Whether to plot city boundary}

\item{test}{logical. (ignored) Only for testing.}
}
\value{
sf
}
\description{
This function returns a polygon of a city boundary based on the UID
}
\examples{
check_urban_boundary(test = TRUE)
}
\references{
Wu, S., Song, Y., An, J. et al. High-resolution greenspace dynamic
data cube from Sentinel-2 satellites over 1028 global major cities.
Sci Data 11, 909 (2024). https://doi.org/10.1038/s41597-024-03746-7
}
