% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{grid_density}
\alias{grid_density}
\title{Generate Density Grid}
\usage{
grid_density(r, grid_size = NULL, stat = c("mean", "sum"), threshould = 0.5)
}
\arguments{
\item{r}{SpatRaster. A single-layer raster representing greenspace change.}

\item{stat}{Character. One of \code{'mean'} or \code{'sum'}.
Determines which summary statistic is returned.}
}
\value{
An \code{sf} object with grid polygons and the requested summary statistic in the column \code{stat_value}.
}
\description{
Divides a greenspace raster into a regular grid and computes
per-cell mean change or point density.
}
\examples{

grid <- grid_density(r)
plot(grid["stat_value"])
}
