% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_gif}
\alias{to_gif}
\title{Convert A Multi-layer Raster to GIF}
\usage{
to_gif(r, fps = 5, width = 600, height = 600, axes = TRUE, title_prefix = NULL)
}
\arguments{
\item{r}{A SpatRaster with multiple layers.}

\item{fps}{numeric. Frames per second (default 5).}

\item{width}{numeric. Width of output GIF in pixels.}

\item{height}{numeric. Height of output GIF in pixels.}

\item{axes}{logical. Draw axes?}

\item{title_prefix}{character or character vector.
Optional prefix or per-frame titles.}
}
\value{
An animated magick image object (GIF).
}
\description{
Export a multi-layer raster (\code{SpatRaster}) to an animated GIF.
}
\examples{
sample_data <- terra::rast(system.file("extdata", "detroit_gs.tif", package = "greenSD"))
gif <- to_gif(sample_data)

}
