% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{vis_change}
\alias{vis_change}
\title{vis_change}
\usage{
vis_change(method = NULL)
}
\arguments{
\item{change_threshould}{description}
}
\value{
list
}
\description{
Statistical analysis of changes of greenspace over years or time during a year.
}
