% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{compute_morphology}
\alias{compute_morphology}
\title{compute_morphology}
\usage{
compute_morphology(p = NULL, directions = 4, grid_size = NULL)
}
\arguments{
\item{p}{SpatRaster or list of SpatRaster.}

\item{directions}{numeric. The number of directions in which patches should be
connected: 4 (default) or 8.}
}
\description{
Compute greenspace morphology metrics at patch or grid level,
including average size (AREA_MN), fragmentation (PD), connectedness (COHESION),
aggregation (AI), and complexity of the shape (SHAPE_AM), related to public health
(Wang et al., 2024)
}
\references{
Wang, H., & Tassinary, L. G. (2024). Association between greenspace morphology
and prevalence of non-communicable diseases mediated by air pollution and physical
activity. Landscape and Urban Planning, 242, 104934.
}
