% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_band_index_by_time}
\alias{get_band_index_by_time}
\title{Get band index based on time period}
\usage{
get_band_index_by_time(time, year)
}
\arguments{
\item{time}{Character vector of length 2. (optional) Start and end dates in \code{"MM-DD"} format
(e.g., \code{c("03-20", "10-15")}). Used to subset the 10-day interval data cube by time.}

\item{year}{numeric. (required) The year of interest.}
}
\description{
A short description...
}
\examples{
get_band_index_by_time(c("03-20", "10-15"), year = 2020)

}
