% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{lc_sem_seg}
\alias{lc_sem_seg}
\title{Classify land cover using multi-source datasets}
\usage{
lc_sem_seg(
  bbox = NULL,
  place = NULL,
  datetime = c(),
  cloud_cover = 10,
  vege_perc = 0,
  zoom = 18,
  sample_size = 5000,
  trian = 0.7
)
}
\arguments{
\item{bbox}{\code{sf}, \code{sfc}, or a numeric vector (xmin, ymin, xmax, ymax)
defining the area of interest. Optional if \code{place} is provided.}

\item{place}{character or vector. (optional) A single line address,
e.g. ("1600 Pennsylvania Ave NW, Washington") or a vector of addresses
(c("Madrid", "Barcelona")).}

\item{datetime}{numeric vector of 2. The time of interest such as
\code{c("2020-08-01", "2020-09-01")}.}

\item{cloud_cover}{numeric. The percentage of cloud coverage.}

\item{vege_perc}{numeric. The percentage of cloud coverage.}

\item{zoom}{numeric. Zoom level of map tile. The default is \code{18}. To
exclude RGB-based map tiles in training random forest model,
use \code{zoom = NULL}.}

\item{sample_size}{numeric. The total number of locations to be sampled
over land cover classes. The default is \code{5000}.}

\item{train}{numeric. The percentage of samples used for training model.
The default is \code{0.7}.}
}
\description{
Generate semantic landcover segmentation based on bands of Sentinel-2-l2a
images by sampling data over the landcover of ESA WorldCover dataset and
using random forest. This function is built based on \code{\link[=get_esa_wc]{get_esa_wc()}} and
the foundation of \code{\link[=get_s2a_ndvi]{get_s2a_ndvi()}}.
}
