% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{sample_values}
\alias{sample_values}
\title{Sample Greenspace Values from Greenspace Seasonality Data Cube or
ESA WorldCover 10m Annual Composites Dataset}
\usage{
sample_values(
  samples = NULL,
  time = NULL,
  source = "gsdc",
  output_bands = NULL,
  cloud_cover = 10,
  vege_perc = 0,
  select = "latest",
  method = "first"
)
}
\arguments{
\item{samples}{A list, matrix, \code{data.frame}, or \code{sf} object of point locations.
Can be a list of length-2 numeric vectors (\code{list(c(lon, lat))}),
a 2-column matrix or data.frame, or an \code{sf} object with POINT geometry in any CRS.}

\item{time}{numeric or vector. The time of interest. See Detail.}

\item{source}{character. The data source for extracting greenspace values:
\code{gsdc} for Greenspace Seasonality Data Cube, \code{esa_ndvi}or \code{esa_landcover} for
ESA WorldCover 10m Annual Dataset, and \code{s2a_ndvi} or \code{s2a_bands} for
Sentinel-2-l2a image data. The default is \code{gsdc}.}
}
\value{
A \code{data.frame} containing greenspace values extracted at each point
across all bands. Each row corresponds to a sample location;
columns represent band values.
}
\description{
Samples values by locatoins from the Greenspace Seasonality Data Cube
developed by Wu et al. (2024) or ESA WorldCover 10m Annual Composites Dataset
by Zanaga et al. (2021).
}
\details{
\code{time}: For the greenspace seasonality data cube, only years from 2019 to 2022
are availabe. For ESA WorldCover 10m Annual Composites Dataset, only 2020
and 2021 are available.

The Greenspace Data Cube is organized into 36 bands per year,
each representing a 10-day interval.
}
\note{
For sampling data from Greenspace Seasonality Data Cube \code{samples} must be
located within the same boundary of an available city in the data cube.
Use \code{\link[=check_available_urban]{check_available_urban()}} and \code{\link[=check_urban_boundary]{check_urban_boundary()}} to see supported
cities and their boundaries.
}
\examples{
# see supported urban areas and their boundaries
check_available_urban()
boundary <- check_urban_boundary(uid = 11)

# sample locations with in the boundary
samples <- sf::st_sample(boundary, size = 20)

# extract values
gs_samples <- sample_values(samples,
                            # time = 2022
                           )

}
\references{
Wu, S., Song, Y., An, J. et al. High-resolution greenspace dynamic
data cube from Sentinel-2 satellites over 1028 global major cities.
Sci Data 11, 909 (2024). https://doi.org/10.1038/s41597-024-03746-7

Zanaga, D., Van De Kerchove, R., De Keersmaecker, W., Souverijns, N.,
Brockmann, C., Quast, R., Wevers, J., Grosu, A., Paccini, A., Vergnaud, S.,
Cartus, O., Santoro, M., Fritz, S., Georgieva, I., Lesiv, M., Carter, S.,
Herold, M., Li, L., Tsendbazar, N.-E., … Arino, O. (2021).
ESA WorldCover 10 m 2020 v100 (Version v100).
Zenodo. https://doi.org/10.5281/zenodo.5571936

Zanaga, D., Van De Kerchove, R., Daems, D., De Keersmaecker, W., Brockmann,
C., Kirches, G., Wevers, J., Cartus, O., Santoro, M., Fritz, S., Lesiv, M.,
Herold, M., Tsendbazar, N.-E., Xu, P., Ramoino, F., & Arino, O. (2022).
ESA WorldCover 10 m 2021 v200 (Version v200).
Zenodo. https://doi.org/10.5281/zenodo.7254221
}
