% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{sample_values}
\alias{sample_values}
\title{Sample Greenspace Values from Data Cube}
\usage{
sample_values(samples = NULL, year = NULL)
}
\arguments{
\item{samples}{A list, matrix, \code{data.frame}, or \code{sf} object of point locations.
Can be a list of length-2 numeric vectors (\code{list(c(lon, lat))}),
a 2-column matrix or data.frame, or an \code{sf} object with POINT geometry in any CRS.}

\item{year}{numeric. The year of interest for the greenspace seasonality data cube.}
}
\value{
A \code{data.frame} containing greenspace values extracted at each point
across all 36 bands. Each row corresponds to a sample location;
columns represent band values.
}
\description{
Samples values by locatoins from the Greenspace Seasonality Data Cube
global dataset developed by Wu et al. (2024).
}
\details{
The Greenspace Data Cube is organized into 36 bands per year,
each representing a 10-day interval.
}
\note{
\code{samples} must be located within the same boundary of an available city in the data cube.
Use \code{\link[=check_available_urban]{check_available_urban()}} and \code{\link[=check_urban_boundary]{check_urban_boundary()}} to see supported
cities and their boundaries.
}
\examples{
# see supported urban areas and their boundaries
check_available_urban()
boundary <- check_urban_boundary(uid = 11)

# sample locations with in the boundary
samples <- sf::st_sample(boundary, size = 20)

# extract values
gs_samples <- sample_values(samples,
                            # year = 2022
                           )

}
\references{
Wu, S., Song, Y., An, J. et al. High-resolution greenspace dynamic
data cube from Sentinel-2 satellites over 1028 global major cities.
Sci Data 11, 909 (2024). https://doi.org/10.1038/s41597-024-03746-7
}
